+++ CHANGES.md
@@ -1,3 +1,8 @@
+#### 1.0.2
+* New: FacetWP support in [grid] shortcode.
+* Move section settings to slide out side panel.
+* Add additional settings to each section.
+
 #### 1.0.1.1
 * Remove unnecessary width declaration on img.